Second Level Update - AI
I consider this to be a major advance since it marks the beginning of the end of the Second Level Development Process.

Short Update Mentions:
 - Changed AIUtils_Fabio.HasVisionOfThePlayer() to ignore the second LayerMask instead of the eigth which was "Glass".
Since the glass is hardly used and the doors (which is on the eigth layer) are way more used, decided to change it.
 - The Player's ability to climb stairs must be revisited since the Player has a hard time to go up the stairs in the Second Level.
Might be the stairs measurements but it can also be the Player's Implementation.

AI Update:
 - Started Second Level AI
 - My basis of the AI is the same as the previously made one, simply made some changes to it.
 - Added my own Behaviours and implemented the majority of them.
 - Created my AIStates as well as AIEvents. Simply added "Fabio_" before any of these names.