Weapon, CharController and Drone Update
Quick Mentions:
 Second Level Update:
 - When the player enters the elevator, the resources are deactivated. This might save us from a few bugs.

Scripts Update:
 - Created DynamicDroneSpawnPoint that will change the Drone's spawn point if needed.
Whenever the drone spawn point is colliding against an object that has a collider, the spawn point automaticly transitions to on top of the player's  head.
This, however, allows a bug on the elevator. If the player jumps and activates the drone, the drone will be out of bounds.
In order to fix this, we can simply lower the closeup Spawn Position of the drone, probably even going inside the player's collider.
More tests need to be done.

 - Implemented ReplenishBullets method on Weapon Script and ReplenishHealth on CharController Script.
This allows all weapons to be filled with bullets as well as the player's health and healthcolor be filled to 100%.
 - Need to add a way to guarantee that there's no enemies active when the player enters the elevator. If not, the enemies will simply pass through the doors.
A possible solution is if there are enemies active, the elevator will not open it's doors. When they're all dead / inactive, open the elevator doors.
Needs more testing

What's missing?
 - Besides everything mentioned previously, I want to find a way to fix the enemies vision. Many times the enemies will see the player twice without losing sight of him and this leads to bugs on their behaviours.
Solution might be the AIManager verify each frame if there are enemies who see the player. If there's even just one, then proceed with the behaviours. If not, start the chase.
This might be a bit heavier on the CPU to verify the enemies "isSeeingPlayer" bool but it just might solve the issue.
Needs more testing.