Second Level Small Update
Quick Mentions:
 DynamicDroneSpawnPosition
 - Fixed CloseSpawnPosition. Doesn't get through the elevator ceiling.

Second Level Update:
 - Fixed enemy's lack of vision if the player was inside the elevator, even if it had the doors open.
 - Elevator is no longer walkable to the enemies.
 - Elevator will no longer be active while all the AI Managers aren't inactive. This also fixes the issue of enemies walking  through the doors of the elevator.

What's missing?
 - As mentioned in the previous update, I want to find a way to fix the enemies vision. Many times the enemies will see the player twice without losing sight of him and this leads to bugs on their behaviours.
Solution might be the AIManager verify each frame if there are enemies who see the player. If there's even just one, then proceed with the behaviours. If not, start the chase.
This might be a bit heavier on the CPU to verify the enemies "isSeeingPlayer" bool but it just might solve the issue.
Needs more testing.
 - The drone, when it activates, has a 90 degree rotation that shouldn't have. Need to fix this.